+++ changelog.md
@@ -1,7 +1,9 @@
 ## Changelog
 
 ### v.1.3.3 -
-- `Fix` Missing friend count in user profile
+- `Fix` Missing friend count in `getUser()`
+- `Fix` Missing user update in `getStat()`
+- `Fix` Summary array in `getStat()`
 
 ### v1.3.2 - Oct 27, 2018
 - `Add` New library logo